feat: Implement detailed cost calculation for Gemini API
This commit introduces a new, detailed cost calculation mechanism for the Gemini API usage, as per the new pricing requirements.

The key changes include:

-   **New CostCalculator Service:** A new service, `CostCalculator`, has been created to encapsulate the complex pricing logic. This service handles different pricing tiers for short and long contexts, as well as for "thinking" mode.

-   **Updated Configuration:** The configuration has been updated to remove the old, single price per token and now includes a comprehensive set of environment variables for different usage types (text, audio, image, video, thinking, long context). All monetary values are handled as high-precision `Decimal` types.

-   **Database Schema Changes:** The database migration has been updated to support the new cost structure. New columns have been added to the `procurement_analyses` table to store the detailed cost breakdown (`cost_input_tokens`, `cost_output_tokens`, `total_cost`, `is_thinking_mode`). The precision of all monetary columns, including existing ones in `procurements`, `donations`, and `budget_ledgers`, has been standardized to `DECIMAL(28, 18)`.

-   **Service and Repository Integration:** The `AnalysisService` and `AnalysisRepository` have been updated to use the new `CostCalculator` and to save the detailed cost information to the database for both new and reused analyses.

-   **Data Models:** All Pydantic models that handle monetary values (`Procurement`, `AnalysisResult`) have been updated to use the `Decimal` type, ensuring consistency throughout the application.

-   **Testing:** New unit tests have been added for the `CostCalculator`, and existing unit and integration tests have been updated to reflect the new changes. All tests, including the E2E test, are passing.

+++ source/public_detective/models/procurements.py
@@ -8,6 +8,7 @@ other related entities.
 """
 
 from datetime import datetime
+from decimal import Decimal
 from enum import IntEnum, StrEnum
 from uuid import UUID
 
@@ -230,7 +231,7 @@ class Procurement(BaseModel):
     object_description: str = Field(..., alias="objetoCompra")
     source_system_link: HttpUrl | str | None = Field(None, alias="linkSistemaOrigem")
     legal_support: LegalSupport = Field(..., alias="amparoLegal")
-    total_awarded_value: float | None = Field(None, alias="valorTotalHomologado")
+    total_awarded_value: Decimal | None = Field(None, alias="valorTotalHomologado")
     is_srp: bool = Field(..., alias="srp")
     government_entity: GovernmentEntity = Field(..., alias="orgaoEntidade")
     procurement_year: int = Field(..., alias="anoCompra")
@@ -243,7 +244,7 @@ class Procurement(BaseModel):
     pncp_control_number: str = Field(..., alias="numeroControlePNCP")
     global_update_date: datetime = Field(..., alias="dataAtualizacaoGlobal")
     dispute_method: DisputeMethod | int = Field(..., alias="modoDisputaId")
-    total_estimated_value: float | None = Field(None, alias="valorTotalEstimado")
+    total_estimated_value: Decimal | None = Field(None, alias="valorTotalEstimado")
     procurement_status: ProcurementStatus | int = Field(..., alias="situacaoCompraId")
     user_name: str = Field(..., alias="usuarioNome")
     electronic_process_link: HttpUrl | str | None = Field(None, alias="linkProcessoEletronico")
